requirements.txt
streamlit==1.50.0
streamlit-option-menu==0.4.0
python-dotenv==1.1.1

google-generativeai==0.8.6

psutil==7.2.2
plotly==6.3.1
python-pptx==1.0.2
reportlab==4.4.4

SpeechRecognition==3.14.4
pyttsx3==2.99
google-genai
streamlit 
python-dotenv

groq
requests
python-dotenv
gtts
requests